Refactor trace writer record node helpers
